% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REGRM.r
\name{REGRM}
\alias{REGRM}
\title{REGRM: Robust estimator for a generalized ratio model
              by iteratively re-weighted least squares (IRLS) algorithm for M-estimation}
\usage{
REGRM(x1, y1, gm = "b", wf = "T", scale = "AAD", c2 = 8,
  rp.max = 100, cg.rt = 0.01)
}
\arguments{
\item{x1}{single explanatory variable}

\item{y1}{objective variable to be imputed}

\item{gm}{indication of gamma value as follows:  \cr
gm="a": gamma=1 \cr
gm="b": gamma=1/2    (conventional ratio model) \cr
gm="c"; gamma=0    (regression model without intercept) \cr}

\item{wf}{weight function (wf=T : Tukey, wf=H : Huber) \cr}

\item{scale}{scale parameter (scale=AAD; scale=MAD)}

\item{c2}{tuning parameter (c2=4, 6, 8) for weight function. Smaller figure is more robust.}

\item{rp.max}{maximum number of iteration}

\item{cg.rt}{convergence condition to stop iteration (default: cg1=0.001)}
}
\value{
a list with the following elements
\describe{
  \item{\code{par}}{robustly estimated ratio of y1 to x1 (beta)}
  \item{\code{res}}{homoscedastic quasi-residuals}
  \item{\code{wt}}{robust weights}
  \item{\code{rp}}{total number of iteration}
  \item{\code{s1}}{changes of the scale (AAD or MAD)}
  \item{\code{efg}}{error flag. 1: acalculia (all weights become zero)  0: successful termination}
}
}
\description{
REGRAM function integrates 12 functions included in RrT.r and RrH.r for easy
             use for users.  Please note that the values for the tuning parameter c2
             allowed in this function is standardized and limited.  If you prefer to use
             other values, please use the functions contained in RrT.r and/or RrH.r directly.
             The turning parameters of those functions are not standardized.
}
\examples{
## Not run:

## End(Not run)
}
